% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_all_listes}
\alias{get_all_listes}
\title{~ api : recuperer toutes les listes sur une thematique}
\usage{
get_all_listes(theme, def_url = "http://referime.aphp.fr/v0.2/")
}
\description{
~ api : recuperer toutes les listes sur une thematique
}
\examples{
\dontrun{
get_all_listes("Chirurgie bariatrique")
}

}
\author{
G. Pressiat
}
