% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_dictionnaire}
\alias{get_dictionnaire}
\title{~ api : recuperer le dictionnaire des listes}
\usage{
get_dictionnaire(def_url = "http://referime.aphp.fr/v0.2/")
}
\description{
~ api : recuperer le dictionnaire des listes
}
\examples{
\dontrun{
get_dictionnaire()
}

}
\author{
G. Pressiat
}
