% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_finess_ej}
\alias{get_finess_ej}
\title{~ api : recuperer les finess juridique Etalab}
\usage{
get_finess_ej(
  departement = c("75", "77", "78", "91", "92", "93", "94", "95"),
  def_url = "http://referime.aphp.fr/v0.1/"
)
}
\description{
~ api : recuperer les finess juridique Etalab
}
\examples{
\dontrun{
get_finess_ej()
get_finess_ej(departement = '75')
get_finess_ej(departement = c('54', '32'))
}
}
\author{
G. Pressiat
}
