% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_liste}
\alias{get_liste}
\title{~ api : recuperer une liste}
\usage{
get_liste(nom_liste, def_url = "http://referime.aphp.fr/v0.2/")
}
\description{
~ api : recuperer une liste
}
\examples{
\dontrun{
get_liste('chir_bariatrique_total')
}

}
\author{
G. Pressiat
}
