% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_table}
\alias{get_table}
\title{~ api : recuperer une table (referentiel)}
\usage{
get_table(table, time_interval = "", def_url = "http://referime.aphp.fr/v0.2/")
}
\description{
~ api : recuperer une table (referentiel)
}
\examples{
\dontrun{
get_table('dictionnaire_tables')
get_table('ccam_actes')
get_table('cim', time_interval = 2018)
get_table('ghm_ghm_regroupement', time_interval = 2017:2018)
}

}
\author{
G. Pressiat
}
