% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_liste}
\alias{tidy_liste}
\title{~ req : transformer une liste en df}
\usage{
tidy_liste(ma_liste, element)
}
\description{
~ req : transformer une liste en df
}
\examples{
\dontrun{
get_liste('chir_bariatrique_total') -> l
tidy_liste(l, "actes")

get_all_listes("Chirurgie bariatrique") -> ll
lapply(ll, function(x)tidy_liste(x, "actes")) \%>\% bind_rows() -> actes_chir_bar
}

}
\author{
G. Pressiat
}
